Move ToValue(..) into its own module.
