+++ graphql-api.cabal
@@ -52,6 +52,7 @@ library
       GraphQL.Resolver
       GraphQL.Value
       GraphQL.Value.FromValue
+      GraphQL.Value.ToValue
   default-language: Haskell2010
 
 test-suite graphql-api-doctests
