+++ src/GraphQL/Value/ToValue.hs
@@ -1,0 +1,58 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DefaultSignatures #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE KindSignatures #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeOperators #-}
+
+module GraphQL.Value.ToValue
+  ( ToValue(..)
+  ) where
+
+import Protolude
+import GraphQL.Value
+import Data.List.NonEmpty (NonEmpty)
+
+-- * ToValue
+
+-- | Turn a Haskell value into a GraphQL value.
+class ToValue a where
+  toValue :: a -> Value' ConstScalar
+
+instance ToValue (Value' ConstScalar) where
+  toValue = identity
+
+-- XXX: Should this just be for Foldable?
+instance ToValue a => ToValue [a] where
+  toValue = toValue . List' . map toValue
+
+instance ToValue a => ToValue (NonEmpty a) where
+  toValue = toValue . makeList
+
+instance ToValue Bool where
+  toValue = ValueBoolean
+
+instance ToValue Int32 where
+  toValue = ValueInt
+
+instance ToValue Double where
+  toValue = ValueFloat
+
+instance ToValue String where
+  toValue = ValueString
+
+-- XXX: Make more generic: any string-like thing rather than just Text.
+instance ToValue Text where
+  toValue = toValue . String
+
+instance ToValue List where
+  toValue = ValueList'
+
+instance ToValue (Object' ConstScalar) where
+  toValue = ValueObject'
+
+
+makeList :: (Functor f, Foldable f, ToValue a) => f a -> List
+makeList = List' . Protolude.toList . map toValue
